+++ src/cdd/constants.py
@@ -6,3 +6,11 @@ ADMINISTRATIVE_LEVEL_TYPE = Choices(
     ("ARRONDISSEMENT", "arrondissement", "Arrondissement"),
     ("VILLAGE", "village", "Village"),
 )
+
+AGENT_ROLE = Choices(
+    ("ACSDCC", "AC-SDCC", "Assistant coordonnateur spécialiste du dcc"),
+    ("FGB", "FGB", "Formateur en gestion à la base"),
+    ("SC", "SC", "Superviseur communal"),
+    ("FT", "FT", "Facilitateur technique"),
+    ("FC", "FC", "Facilitateur communautaire"),
+)