+++ src/cdd/settings.py
@@ -38,6 +38,7 @@ SECRET_KEY = env("SECRET_KEY")
 
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = env.bool("DEBUG", False)
+print(DEBUG)
 
 ALLOWED_HOSTS += ["127.0.0.1"]