+++ src/CHANGELOG.md
@@ -4,6 +4,9 @@ List of features add to the project
 
 ## unreleased
 
+- change administrative level allocation to agent
+- `facilitator` is now agent with different role. `facilitator` is now an `agent` `role`
+- add `role` field to `facilitator` model
 - updating `views` relatives to phases, activities and tasks
 - updated `FacilitatorListTableView` for search filter
 - updated `FacilitatorListView` to match new administrative level structure